train/requirements.txt
mlflow==2.14.0
pandas
autogluon.timeseries
scikit-learn
numpy
pyarrow
fastparquet
kafka-python
